# src/swagger.yaml
swagger: '2.0'
info:
  description: >-
    This is the Moscord Ecommerce Service.
  version: 0.0.1
  title: Moscord Ecommerce Swagger
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: halili.aubrian@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
basePath: /api
...
schemes:
- "https"
- "http"
paths:
...
  /sellers:
    get:
      tags:
      - "sellers"
      summary: "Return seller list"
      description: ""
      produces:
      - "application/json"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Seller"
        400:
          description: "Invalid input"
    post:
      tags:
      - "sellers"
      summary: "Create seller"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Seller object"
        required: true
        schema:
          $ref: "#/definitions/Seller"
      responses:
        400:
          description: "Invalid input"
  /sellers/{id}:
    get:
      tags:
      - "sellers"
      summary: "Return seller object"
      description: ""
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Seller id"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Seller"
    put:
      tags:
      - "sellers"
      summary: "Updated seller"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Seller id"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Seller object"
        required: true
        schema:
          $ref: "#/definitions/Seller"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Seller"
        404:
          description: "Not found"
    delete:
      tags:
      - "sellers"
      summary: "Delete seller"
      description: ""
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "seller id"
        required: true
        type: "string"
      responses:
        404:
          description: "Not found"
...
definitions:
...
  Seller:
    type: "object"
    required:
    - "name"
    - "email"
    - "date_of_birth"
    - "phone"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      date_of_birth:
        type: "date"
      phone:
        type: "number"
...